# backend/api.yaml
openapi: 3.0.3
info:
  version: 0.1.0
  title: backend api
servers:
  - url: http://localhost:3000
    description: servidor de desenvolvimento
paths:
  /reservas:
    get:
      summary: Lista reservas
      responses:
        "200":
          description: lista com as reservas feitas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reserva"
        "500":
          description: erro ao listar reservas
  /reserva/{id}:
    get:
      summary: Detalhes da reserva por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1
          description: id da reserva
      responses:
        "200":
          description: detalhes da reserva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserva"
        "400":
          description: reserva não encontrada
        "500":
          description: erro ao buscar reserva
    put:
      summary: Atualiza informações de uma reserva
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1
          description: id da reserva
      requestBody:
        $ref: "#/components/requestBodies/NovaReservaBody"
      responses:
        "200":
          description: alterações salvas
          content:
            appplication/json:
              schema:
                $ref: "#/components/schemas/Reserva"
        "500":
          description: erro ao alterar reserva
    delete:
      summary: Remove uma reserva
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1
          description: id da reserva
      responses:
        "200":
          description: reserva removida com sucesso
        "500":
          description: erro ao remover reserva
  /reserva:
    post:
      summary: Adiciona nova reserva
      requestBody:
        $ref: "#/components/requestBodies/NovaReservaBody"
      responses:
        "200":
          description: nova reserva feitas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserva"
        "500":
          description: erro ao criar reserva
  /destinos:
    get:
      summary: Lista destinos
      responses:
        "200":
          description: lista com todos os destinos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destino"
  /destino/{id}:
    get:
      summary: Detalhes do destino por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1
          description: id do destino
      responses:
        "200":
          description: detalhes do destino
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destino"
        "400":
          description: destino não encontrado
        "500":
          description: erro ao buscar destino
components:
  schemas:
    Reserva:
      type: object
      properties:
        partida:
          type: string
          example: "2022-08-20T00:00:00.000Z"
        retorno:
          type: string
          example: "2022-08-30T00:00:00.000Z"
        destino:
          $ref: "#/components/schemas/Destino"
        passageiro:
          $ref: "#/components/schemas/Passageiro"
        preco:
          type: number
          example: 500
    destino:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Fortaleza
        preco:
          type: integer
          example: 500
    passageiro:
      type: object
      properties:
        nome:
          type: string
          example: João da Silva
        cpf:
          type: string
          example: 000.000.000-00
  requestBodies:
    NovaReservaBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              id_destino:
                type: integer
              partida:
                type: string
              retorno:
                type: string
              passageiro:
                $ref: "#/components/schemas/Passageiro"
          example:
            id_destino: 1
            partida: "2022-01-25"
            retorno: "2022-01-25"
            passageiro:
              nome: João da Silva
              cpf: 000.000.000-00